Test folder renamed from "bench" to "automated"
