+++ test/automated/automated_init.rb
