+++ test/automated/pure_virtual_method_is_implemented_in_subclass.rb
@@ -1,4 +1,4 @@
-require_relative './bench_init'
+require_relative './automated_init'
 
 module Fixture
   module ImplementedInSubclass
